Huami: show event title for events as reminder
The Mi app only has a reminder field and no description field, so
I'm assuming we can only send one string.

In bffb8e8f876186fe61e7ce73da2b40d89c948e91 it was noticed that the
title was not displaying so I'm changing this for all Huami devices.