Amazfit GTS2/GTR2: Fix setting menu items
At least on the GTR2e, including the "watchface" menu item causes the
configuration request to be ignored.  Assuming the other GTS2 and GTR2
variants behave the same, this patch ensures the forceWatchface argument
is set to false when setting display items.

Additionally, we now use the correct resource ID for the default
settings.