Add support for alarm notifications from Google Clock
The Google Clock from the Play Store (`com.google.android.deskclock`) is
in more strict in how broadcasts are sent to the system, as they require
a permission that is restricted to `signature|system`. This requires us
to process the notification from the app in order to determine if an
alarm is ongoing, snoozed or dismissed.