huawei: Improve user feedback in error cases
Currently the user isn't informed about errors when connecting to
the watch and is left with a working BT connection that isn't used at all.

Add toasts when the HiChainRequest fails or times out and disconnect
the phone. Without a successful HiChain established the connection is
useless anyways and it causes the phone to be not discoverable any more.

In addition add a timeout to the HiChainRequest, one longer for the first
pairing, where the user needs to confirm the pairing request on the watch.
The short delay is used for subsequent HiChainRequests.

The watch might not answer HiChainRequests when it was paired with a
different phone, so the added timeout and toast improves user experience
a lot since it's now clear that there was a problem.

Related to issue #4148
Related to issue #4061
Fixes issue #4062

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>