HuaweiSampleProvider: Improve getRawOrderedActivitySamples
Currently only ActivitySamples that *start* within the specified time
frame are loaded from the DB, but ActivitySamples that end in the specified
time frame are ignored. On the Dashboard and the activity summary chart
some activities might thus be missing.

Since the sleep activity usually starts hours before midnight it doesn't
get loaded and the Dashboard doesn't properly display the sleep phase.

To fix this issue load all ActivitySamples that overlay the specified
time frame, that means it also loads samples that end within the timeframe.

Fixes issue #4510

Signed-off-by: Patrick Rudolph <rudolphpatrick05@gmail.com>